refactor: extract run summary output helpers to output.run_summary
